fix: update amino data length checks to use uint64 for consistency(v1.7.0) (#3341)
* fix: update data length checks to use uint64 for consistency

* fix: improve error handling for transaction data unmarshalling

+++ libs/tendermint/types/part_set.go
@@ -53,7 +53,7 @@ func (part *Part) UnmarshalFromAmino(cdc *amino.Codec, data []byte) error {
 			}
 
 			data = data[n:]
-			if len(data) < int(dataLen) {
+			if uint64(len(data)) < dataLen {
 				return fmt.Errorf("not enough data for %s, need %d, have %d", aminoType, dataLen, len(data))
 			}
 			subData = data[:dataLen]
@@ -154,7 +154,7 @@ func (psh *PartSetHeader) UnmarshalFromAmino(_ *amino.Codec, data []byte) error 
 			}
 
 			data = data[n:]
-			if len(data) < int(dataLen) {
+			if uint64(len(data)) < dataLen {
 				return fmt.Errorf("not enough data for %s, need %d, have %d", aminoType, dataLen, len(data))
 			}
 			subData = data[:dataLen]
